@@ -13,18 +12,18 @@
         "e2e": "ng e2e"
     },
     "dependencies": {
-        "@angular/animations": "11.2.10",
-        "@angular/cdk": "11.2.9",
-        "@angular/common": "11.2.10",
-        "@angular/compiler": "11.2.10",
-        "@angular/core": "11.2.10",
-        "@angular/forms": "11.2.10",
-        "@angular/material": "11.2.9",
-        "@angular/material-moment-adapter": "11.2.9",
-        "@angular/platform-browser": "11.2.10",
-        "@angular/platform-browser-dynamic": "11.2.10",
-        "@angular/router": "11.2.10",
-        "@angular/service-worker": "11.2.10",
+        "@angular/animations": "12.0.0",
+        "@angular/cdk": "12.0.0",
+        "@angular/common": "12.0.0",
+        "@angular/compiler": "12.0.0",
+        "@angular/core": "12.0.0",
+        "@angular/forms": "12.0.0",
+        "@angular/material": "12.0.0",
+        "@angular/material-moment-adapter": "12.0.0",
+        "@angular/platform-browser": "12.0.0",
+        "@angular/platform-browser-dynamic": "12.0.0",
+        "@angular/router": "12.0.0",
+        "@angular/service-worker": "12.0.0",
         "@fullcalendar/angular": "4.4.5-beta",
         "@fullcalendar/core": "4.4.2",
         "@fullcalendar/daygrid": "4.4.2",
@@ -33,55 +32,63 @@
         "@fullcalendar/moment": "4.4.2",
         "@fullcalendar/rrule": "4.4.2",
         "@fullcalendar/timegrid": "4.4.2",
+        "@auth0/angular-jwt": "5.0.2",
         "@microsoft/signalr": "5.0.4",
-        "apexcharts": "3.26.0",
-        "aws-amplify": "3.3.26",
+        "apexcharts": "3.26.2",
         "crypto-js": "3.3.0",
         "highlight.js": "10.7.2",
         "lodash-es": "4.17.21",
         "moment": "2.29.1",
         "ng-apexcharts": "1.5.9",
-        "ngx-markdown": "11.1.2",
-        "ngx-quill": "13.2.0",
-        "perfect-scrollbar": "1.5.0",
+        "ngx-markdown": "11.1.3",
+        "ngx-quill": "14.0.0",
+        "perfect-scrollbar": "1.5.1",
         "quill": "1.3.7",
         "rrule": "2.6.8",
         "rxjs": "6.6.7",
-        "tslib": "2.1.0",
+        "tslib": "2.2.0",
         "web-animations-js": "2.3.2",
         "zone.js": "0.11.4"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "0.1102.9",
-        "@angular/cli": "11.2.9",
-        "@angular/compiler-cli": "11.2.10",
-        "@angular/language-service": "11.2.10",
+        "@angular-devkit/build-angular": "12.0.0",
+        "@angular-eslint/builder": "12.0.0",
+        "@angular-eslint/eslint-plugin": "12.0.0",
+        "@angular-eslint/eslint-plugin-template": "12.0.0",
+        "@angular-eslint/schematics": "12.0.0",
+        "@angular-eslint/template-parser": "12.0.0",
+        "@angular/cli": "12.0.0",
+        "@angular/compiler-cli": "12.0.0",
+        "@angular/language-service": "12.0.0",
         "@tailwindcss/aspect-ratio": "0.2.0",
         "@tailwindcss/line-clamp": "0.2.0",
         "@tailwindcss/typography": "0.4.0",
         "@types/chroma-js": "2.1.3",
         "@types/crypto-js": "3.1.47",
         "@types/highlight.js": "10.1.0",
-        "@types/jasmine": "3.6.8",
-        "@types/lodash": "4.14.168",
+        "@types/jasmine": "3.6.11",
+        "@types/lodash": "4.14.169",
         "@types/lodash-es": "4.17.4",
-        "@types/node": "12.20.6",
+        "@types/node": "12.20.13",
+        "@typescript-eslint/eslint-plugin": "4.23.0",
+        "@typescript-eslint/parser": "4.23.0",
         "autoprefixer": "10.2.5",
         "chroma-js": "2.1.1",
-        "codelyzer": "6.0.1",
-        "jasmine-core": "3.6.0",
+        "eslint": "7.26.0",
+        "eslint-plugin-import": "2.23.0",
+        "eslint-plugin-jsdoc": "34.2.2",
+        "eslint-plugin-prefer-arrow": "1.2.3",
+        "jasmine-core": "3.7.1",
         "jasmine-spec-reporter": "5.0.2",
-        "karma": "6.1.2",
+        "karma": "6.3.2",
         "karma-chrome-launcher": "3.1.0",
         "karma-coverage": "2.0.3",
         "karma-jasmine": "4.0.1",
-        "karma-jasmine-html-reporter": "1.5.4",
+        "karma-jasmine-html-reporter": "1.6.0",
         "lodash": "4.17.21",
-        "postcss": "8.2.10",
+        "postcss": "8.2.15",
         "protractor": "7.0.0",
-        "tailwindcss": "2.1.1",
-        "ts-node": "8.3.0",
-        "tslint": "6.1.3",
-        "typescript": "4.1.5"
+        "tailwindcss": "2.1.2",
+        "typescript": "4.2.4"
     }
 }
