@@ -1,4 +1,4 @@
-/* tslint:disable:max-line-length */
+/* eslint-disable */
 export const contacts = [
     {
         id          : 'cd5fa417-b667-482d-b208-798d9da3213c',
