@@ -1,6 +1,6 @@
 import { Shortcut } from 'app/ui/layout/common/shortcuts/shortcuts.types';
 
-/* tslint:disable:max-line-length */
+/* eslint-disable */
 export const shortcuts: Shortcut[] = [
     {
         id         : 'a1ae91d3-e2cb-459b-9be9-a184694f548b',
