@@ -1,4 +1,4 @@
-/* tslint:disable:max-line-length */
+/* eslint-disable */
 export const user: any = {
     id    : 'cfaad35d-07a3-4447-a6c3-d8c3d54fd5df',
     name  : 'Brian Hughes',
