@@ -4,7 +4,7 @@ import { FormGroup } from '@angular/forms';
 import { BehaviorSubject, Subject } from 'rxjs';
 import { filter, first, map, switchMap, takeUntil, tap, withLatestFrom } from 'rxjs/operators';
 import { MatDialog } from '@angular/material/dialog';
-import { FuseAnimations } from '@fuse/animations';
+import { fuseAnimations } from '@fuse/animations';
 
 import { Profile, ProfileConnectionInfo, ProfileGeneralInfo, ProfilePersonalInfo } from '../../profile.model';
 import { ProfileService } from '../../_services/profile.service';
@@ -19,7 +19,7 @@ import { ActivatedRoute } from '@angular/router';
     selector       : 'profile-main',
     templateUrl    : './profile-about.component.html',
     encapsulation  : ViewEncapsulation.None,
-    animations   : FuseAnimations
+    animations   : fuseAnimations
 })
 export class ProfileAboutComponent implements OnInit, OnDestroy
 {
