@@ -18,7 +18,7 @@ import { FormControl } from '@angular/forms';
 import { HttpClient } from '@angular/common/http';
 import { of, Subject } from 'rxjs';
 import { catchError, debounceTime, filter, finalize, map, switchMap, takeUntil, tap } from 'rxjs/operators';
-import { FuseAnimations } from '@fuse/animations/public-api';
+import { fuseAnimations } from '@fuse/animations/public-api';
 import { PersonAutocompletes } from '@ui/layout/common/search/search-bar.models';
 import { Observable } from 'rxjs/internal/Observable';
 import { ApiResponse } from '@shared/shared.models';
@@ -30,7 +30,7 @@ import { Environment } from '@shared/environment.model';
     templateUrl  : './search.component.html',
     encapsulation: ViewEncapsulation.None,
     exportAs     : 'fuseSearch',
-    animations   : FuseAnimations
+    animations   : fuseAnimations
 })
 export class SearchComponent implements OnChanges, OnInit, OnDestroy
 {
